Refactor ORM to use Where for partition
Removes the Partition() method from QueryBuilder. Partition keys
must now be provided via Where().Equals(). Internal query logic
validates that the partition key is present.

Also changes Updated/Created fields from int64 to int32 and updates
frontend types to match backend PascalCase naming.

+++ backend/handlers/productos-indexer.go
@@ -91,7 +91,7 @@ func GetProductsIndexDelta(req *core.HandlerArgs) core.HandlerResponse {
 		e := &productosDelta[index]
 		marcasCategoriasIDs.AddIf(e.MarcaID)
 		marcasCategoriasIDs.AddIfBulk(e.CategoriasIDs...)
-		e.Updated = int64(e.NameUpdated)
+		e.Updated = e.NameUpdated
 		e.NameUpdated = 0
 	}
